fix(button): better handle synchronous resets
When the button has no timeouts we were borking on the `reset` function because we thought we were still processing. This was because `reset` was landing before `setState` callback was. This patch also better handles when the component is unmounted. It will no longer update the state butcontinue to callback any attached handlers.